Filter Panel: ColorPicker Eyedropper Updated
See: #9465
See also: #9488
